fix(scripts): introduce getBlockRange and standardize to [start,end) block ranges (#141)
As the title says. 

This fixes subtle off-by-one errors we had due to HyperSync / RPC node
not treating the "to block" the same way (exclusive vs inclusive).
Also `startTimestamp` is now guaranteed to be included and
`endTimestamp` excluded.

The main changes are in `events.ts` and its unit tests. The rest is
adapting existing code to call the new method.

Part of ENG-365