fix: use end timestamp for fetching referrals (#155)
As the title - fetch referrals currently returns all referrals until the
time of script execution. This is unnecessary, because we only need to
calculate rewards for referrals up to the reward end period. It also
causes extra confusion when peer reviewing the reward distribution,
since the output of the script is time dependent.

Also fixes referral buffer time incorrect conversion.