refactor: rename `calculateRevenue` -> `calculateKpi` and related code (#171)
As the title says.

No behavior was changed. Except for generated file names which are now
`kpi.csv` instead of `revenue.csv`.
All scripts were updated accordingly.

In the implementations of `calculateKpi` I've usually left variables or
internal functions called `revenue` since for most of these the KPI is
really a "revenue".

Part of ENG-394