feat(contracts): Implement initial Data Availability contract (#111)
For context on the full flow, see [this
writeup](https://github.com/divvi-xyz/risc0-example/pull/3/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R6).

Adds the initial version of the Data Availability contract. As a
best-practice, this contract does not actually store the data, but
instead emits events for each "upload" such that the history can be
fully re-created via indexing. Instead of storing the data, the contract
maintains a streaming hash of the data at each timestamp by XORing the
hash of each user:value pair. Since XOR is commutative, this allows an
uploader to upload data for a particular timestamp in any order, and the
final hash will always be the same.

The verification process will look as follows:
* In a trusted zkVM, calculations will be performed on raw data in order
to generate the list of user:value pairs stored in the contract.
* The zkVM will compute the hash of these user:value pairs in the same
way the contract does.
* The zkVM generates a proof of this computation, and attaches the hash
to the journal.
* This proof is uploaded to the contract along with the hash from the
trusted compute environment
* The contract verifies the proof, and compares the journal hash to the
contract hash for the given timestamp.
* If they match, we can be confident that the trusted code running in
the zkVM produced the same data as those previously streamed to the
contract, and can mark the contract data as verified.