chore: ensure delays to avoid rate limits in calculateKpi (#178)
When running the calculateKpi scripts on CI, it was the _getNearestBlock
function that started failing. On closer inspection, it seems envio has
no rate limits but DefiLlama has a rate limit of 500 requests/min. In
the worst case, we call _getNearestBlock for every user in the reward
period if they were referred within the period (i.e. we need to get the
block corresponding to their referral)

This PR adds:
- better error messaging on _getNearestBlock (the error response was
previously logged as `[object Object]`
- delay after every kpi batch transaction to keep requests under the
rate limit

Related to ENG-392