feat: improve DefiLlama rate limit handling (#220)
This PR adds the `bottleneck` npm package to manage the DefiLlama rate
limit at the api request level. This is much more efficient than
managing at the kpi calc level, since some calculations don't require a
request to DefiLlama.

Example of a
[job](https://github.com/divvi-xyz/divvi-protocol-v0/actions/runs/15451499250/job/43494472674?pr=220)
running with this change, which took ~15 minutes. It ran the same logic
as the other jobs in the CI which take ~90 minutes.

With this change, I could also change the frequency of the scheduled
calc and upload job back to 1 hour.

Related to ENG-414