feat: use Redis to cache block number (#222)
This PR allows the fetchReferrals and calculateKpi scripts to take
optional redis configs, and applies a redis cache to the DefiLlama
requests. This means that across calculations, we will no longer need to
hit DefiLlama to get the block number if it was computed in a previous
iteration. This should reduce computation and speed up the calculation.

Related to ENG-414