ci: generate docs (#228)
When changes are needed to the generated docs, the workflow
automatically commits them and pushes to the branch.

We're using a custom PAT (Divvi Bot) because the default one causes the
workflow to be skipped once the "docs: auto-update" commit has been
pushed.
This is GitHub Action's loop protection.

I've added a check to ensure the new `docs` step won't run into a loop,
by checking the last commit message.
Though `yarn docs` should always be the same, but just in case there's a
problem with that.

Note: only review the workflow changes, the docs part was generated.

---------

Co-authored-by: Divvi Bot <89419329+divvi-bot@users.noreply.github.com>