feat: Add segmentedKpi field and add to scout game reward output (#225)
`segmentedKpi` field added to KPI output which is a map from string to
numbers, so any wanted breakdown of KPI can be included. Then the
specific rewards processing file for the campaign can be updated to
process this and include it in the rewards.json output to be used in
analysis/dashboard. KPI breakdown by blockchain added for scout-game
campaign.

**Testing**
Ran locally and rewards.json output was:
```
[
  {
    "referrerId": "0x5e23d5be257d9140d4c5b12654111a4d4e18d9b2",
    "kpi": "12",
    "rewardAmount": "174846083434101646003",
    "baseKpi": "0",
    "celoKpi": "12",
    "polygonKpi": "0"
  },
  {
    "referrerId": "0x124d8fad33e0b9fe9f3e1e90d0fc0055abe8ca8d",
    "kpi": "1",
    "rewardAmount": "50473716668715175624",
    "baseKpi": "0",
    "celoKpi": "1",
    "polygonKpi": "0"
  },
  {
    "referrerId": "0xe6e6c0ddf053822bbbc1f39218f30ea25fca58dd",
    "kpi": "50948",
    "rewardAmount": "11392757558913212937413",
    "baseKpi": "45973",
    "celoKpi": "4568",
    "polygonKpi": "407"
  },
  ...
]
```

---------

Co-authored-by: Divvi Bot <89419329+divvi-bot@users.noreply.github.com>