chore: add slack notification on kpi upload CI failure (#255)
This PR adds a slack notification to the on-call slack channel when the
periodic kpi calculation and upload CI fails. I've been keeping an eye
on this workflow for the last few weeks and it is quite stable, so I
don't expect this alert to fire often.

I followed the integration steps
[here](https://tools.slack.dev/slack-github-action/sending-techniques/sending-data-slack-incoming-webhook/post-inline-text-message).
Example of slack alert is
[here](https://valora-app.slack.com/archives/C02N3AR2P2S/p1750245850795729).

Fixes ENG-392