fix: Use correct endTimestamp in uploadCurrentPeriodKpis script for reward calculation (#253)
Found this when looking at new scout game dashboard cycle and the reward
amounts seemed really small.

Was just using the `endTimestampExclusive` to pass to reward function,
which is the start of the current hour (or passed in timestamp) which is
incorrect. Should instead be passing in the endTimestamp of the
campaign. This is because in some scripts the overall amount to reward
is calculated based on the end timestamp (like scout game), so was
incorrect.