feat: add function to get rewards exclude list (#312)
This PR facilitates Valora / VEarn's integration with Divvi by making it
ineligible for rewards:
- turn the ofac sanctioned addresses script into a function that returns
a list of addresses that are excluded from divvi rewards. currently,
this should be OFAC sanctioned addresses + the valora entities.
- there are some addresses that we never want to reward, e.g. ofac
sanctioned addresses and for now, valora. this will be solved by a
campaign specific inclusion list (e.g. Karma Gap) but when the campaign
has none, we should use this function and write the results to the
rewards directory. in this PR, the celo-pg and lisk campaigns have been
updated to use this function.
- remove the `failOnExclude` behaviour and replace with logging, to
remove manual actions that block automating the reward distribution
process.

The Dune dashboard displays campaign data using the rewards.json, which
means that we should _not_ filter kpi data for excluded referrers.
Instead, we should only discount the kpi data in the final reward
calculation.


Fixes ENG-504