feat: Add mantle kpi/reward calculation (#330)
For
[ENG-532](https://linear.app/valora/issue/ENG-532/mantle-kpirewards-pipeline).
Once this is merged, can use as the reward function ID to [deploy the
reward pool](https://linear.app/valora/issue/ENG-530/reward-pool). Then
can fill in the reward pool address, and once we deposit into the pool
and have the #MNT amount, can update the reward amount value.

Still need to add the mantle campaign entity to REWARDS_PROVIDERS to be
able to fetch referrals, is this entity made?

---------

Co-authored-by: Divvi Bot <89419329+divvi-bot@users.noreply.github.com>