refactor: allow kpi handlers to return multiple kpis per user (#327)
Currently, the calculateKpi handlers take a single user address and
returns a single kpi value for the reward period. the kpi value is then
attributed to their referrer (from `fetchReferrals`).

For the tether campaign experiment, we want to be able to attribute each
transaction to the referrer in the referral tag on the tx, which
requires calculateKpi to be able to return multiple kpi values per end
user (one kpi value per referrer).

This PR updates the calculateKpi script and types to handle multiple
kpis per end user. This discriminated union approach minimises the
changes needed to the existing scripts.

Related to ENG-518