feat(contracts): Add RISC0 verification to Data Availability contract (#120)
Updates the `DataAvailability` contract to support verification using
RISC0's on-chain proof verification system. See
[here](https://github.com/divvi-xyz/risc0-example/pull/3/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R6)
for a more in-depth explanation of how this fits together with the rest
of the PRs!

The diff of this PR looks big, but the majority is just copy-pasting
RISC0 contract definitions that are required in order for us to add
verification.