chore: add get referrer from tx logic for tether campaign (#338)
This PR adds the logic for getting the referrer from the transaction. I
took the code from the `submitReferral` endpoint and tried to preserve
as much as possible while removing the offchain portion and support for
tag format 0. for simplicity, the Tether campaign will expect builders
to use v2 of the referral sdk. We can add these files to a shared module
eventually.

To make the review easier, here are the files that are added:
- `alignMagicPrefixAndPadHex.ts` - copied from
[here](https://github.com/divvi-xyz/divvi-relayer/blob/main/src/endpoints/submitReferral/alignMagicPrefixAndPadHex.ts)
with no changes
- `getTransactionInfo.ts` - copied from
[here](https://github.com/divvi-xyz/divvi-relayer/blob/main/src/endpoints/submitReferral/getTransactionInfo.ts)
with a change to move `TransactionInfo` inside the file and turn the
http errors into standard errors
- `getUserOperations.ts` - copied from
[here](https://github.com/divvi-xyz/divvi-relayer/blob/main/src/endpoints/submitReferral/getUserOperations.ts)
with no changes
- `parseReferral.ts` - copied from
[here](https://github.com/divvi-xyz/divvi-relayer/blob/main/src/endpoints/submitReferral/parseReferral.ts)
with changes to remove logic for offchain path and format tag v0 path
- the changes in `getReferrerIdFromTx.ts` mostly follows the logic from
[here](https://github.com/divvi-xyz/divvi-relayer/blob/029088c86b39c2b3b5d6032dead08833ce74ed05/src/endpoints/submitReferral/index.ts#L260-L356),
minus the support for offchain and tag format v0

Related to ENG-518

---------

Co-authored-by: Divvi Bot <89419329+divvi-bot@users.noreply.github.com>