fix: ensure supported networks have viem client config (#356)
I forgot to add viem client config for the new networks supported by the
Tether campaign. I expected typescript to help us with where to add
config, but this `Partial` type got in the way. This PR removes
`Partial` as much as possible, and adds the missing viem client config.