fix: exclude unregistered builders from kpi calcs for tether campaign (#357)
Since we get the referrers from the transactions for this Tether
campaign, it is possible that there are referral tags added by builders
who are not registered for the campaign. This PR adds logic at the end
of the calculation to ignore any kpis from builders who are not
registered at the time of calculation.

Related to ENG-518

---------

Co-authored-by: Divvi Bot <89419329+divvi-bot@users.noreply.github.com>
Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>