feat(protocol): Add protocol fee to RewardPool (#407)
For
[ENG-578](https://linear.app/valora/issue/ENG-578/update-rewardpoolfactory-to-set-default-protocol-fee-and-reserve)
and
[ENG-579](https://linear.app/valora/issue/ENG-579/update-rewardpool-contract-to-collect-fees-when-adding-rewards).

Adds a `protocolFee` and `reserveAddress` to the `RewardPool` contract.
When allocating rewards to users, sends a protocol fee to the reserve
address _on top_ of the user's reward, such that users always get
rewarded the specified amount. The `RewardPoolFactory` contract has been
updated to allow us to set the `reserveAddress` and `protocolFee` that
are set when deploying a new `RewardPool` contract; these parameters are
pulled _from the factory state_, and are not parameters to
`createRewardPool` so that end users cannot override them.

Since the `owner` on a `RewardPool` has the ability to update
`protocolFee` and `reserveAddress` for a given reward pool, I've updated
the `owner` address that was previously passed in to `createRewardPool`
on the factory to follow the same pattern as `protocolFee` and
`reserveAddress`; it can now be set by the factory admin as a piece of
state on the factory, and is passed into the `RewardPool` init method.
In a self-service deployment model, this prevents campaign managers from
overriding the owner and updating the `protocolFee` or `reserveAddress`
on the deployed reward pool.