feat: Create linear and sqrt reward libraries for onchain reward calculation (#430)
Adds the first two reward function libraries for on chain reward
calculation.

To use in the reward contract should have the reward pool take a
contract address as an input param (the address of the deployed library
to use), and use `delegatecall` passing the function name
`calculateReward` and the expected input (list of (kpi, referrerAddress)
pairs, and a totalRewardAmount).

Future reward functions added should follow the same pattern of creating
a new library and having the same function name and input params. This
way we don't need to upgrade/redeploy any of the reward pool stuff, we
just deploy the new reward library and use the address in the reward
pool.

---------

Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>