chore: Add metadata for RewardPoolFactory upgrades (#485)
Adds in deployment/upgrade metadata for new `RewardPoolFactory`
contracts. The goal here was to upgrade the `RewardPoolFactory`
contracts on all chains we have them currently deployed, as well as set
the `RewardPool` implementation to the latest version.

I wrote up a runbook for this process
[here](https://www.notion.so/valora-inc/RewardPoolFactory-Upgrade-Process-2656b927f5ac80618b5dffb193e336a9?showMoveTo=true&saveParent=true)
and we were able to complete the full process for Celo, Base, Ethereum,
and Optimism. We were able to complete the process partially for Morph
but were unable to perform the upgrade due to errors from Defender.
Further Defender errors prevented us from deploying anything to Mantle.

I've also deployed the `RewardFunction` implementation contracts to each
chain, and documented their deploy addresses. For bookkeeping, I've
started to keep track of both the `RewardPoolFactory` implementation
commit, as well as the commit of the `RewardPool` implementation set on
each `RewardPoolFactory`. We should try to keep these up to date.

---------

Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>